Extend comment-based testing framework with type checker support
