+++ eslint-bridge/tests/linting/eslint/rules/comment-based/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "target": "es2018",
+    "allowJs": true,
+  },
+  "include": ["*.js", "*.ts"]
+}
